Create README.md
I am pleased to share with you all that I have completed the given task for my internship at The Sparks Foundation. This was the task to make a Machine Learning model, that creates a Decision Tree classifier and visualize it graphically and then feed any new data to this classifier, it would be able to predict the right class accordingly.